% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{format_similar_websites}
\alias{format_similar_websites}
\title{Format similar websites response}
\usage{
format_similar_websites(response)
}
\arguments{
\item{response}{the \code{curl} object recieved as a response from \code{get_similar_websites}}
}
\value{
\code{data.frame} with domain, status_code, similar websites list, category, category rank and timestamp
}
\description{
Format similar websites response
}
\examples{
\dontrun{
format_visits(response =
  get_visits("my_email", token = "my_token", domain = "facebook.com"))
}
}
