% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{format_visits}
\alias{format_visits}
\title{Format visits response}
\usage{
format_visits(response)
}
\arguments{
\item{response}{the \code{curl} object recieved as a response from \code{get_visits}}
}
\value{
\code{data.frame} with status, domain, date, visits and timestamp
}
\description{
Format visits response
}
\examples{
\dontrun{
format_visits(response =
  get_visits("my_email", token = "my_token", domain = "facebook.com"))
}
}
