% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_category_rank}
\alias{get_category_rank}
\title{Get category rank}
\usage{
get_category_rank(api_key, domain)
}
\arguments{
\item{api_key}{similarweb API key}

\item{domain}{domain of interest, do not include schemas like \code{www.} or \code{https:}}
}
\value{
The \code{curl} object with \code{category, rank}
}
\description{
Get category rank
}
\examples{
\dontrun{
get_category_rank(api_key = "my_api_key", domain = "facebook.com")
}
}
