% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_rank_and_reach}
\alias{get_rank_and_reach}
\title{Get rank and reach}
\usage{
get_rank_and_reach(api_key, domain)
}
\arguments{
\item{api_key}{similarweb API key}

\item{domain}{domain of interest, do not include schemas like \code{www.} or \code{https:}}
}
\value{
The \code{curl} object with overall \code{rank} and reach information
}
\description{
Get rank and reach
}
\examples{
\dontrun{
get_rank_and_reach(api_key = "my_api_key", domain = "google.com")
}
}
