% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_similar_websites}
\alias{get_similar_websites}
\title{Get similar websites}
\usage{
get_similar_websites(api_key, domain, start_month, end_month)
}
\arguments{
\item{api_key}{similarweb API key}

\item{domain}{domain of interest, do not include schemas like \code{www.} or \code{https:}}

\item{start_month}{date in the format YYYY-MM}

\item{end_month}{date in the format YYYY-MM}
}
\value{
The \code{curl} object with overall 40 most similar sites, global ranking, category, and category ranking
}
\description{
Get similar websites
}
\examples{
\dontrun{
get_similar_websites(api_key = "my_api_key",
                     domain = "coinmarketcap.com",
                     start_month = "2018-01",
                     end_month = "2018-02")
}
}
