% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_visits}
\alias{get_visits}
\title{Get domain visits}
\usage{
get_visits(api_key, domain, start, end, gr, md = "False")
}
\arguments{
\item{api_key}{similarweb API key}

\item{domain}{domain of interest, do not include schemas like \code{www.} or \code{https:}}

\item{start}{start date \code{MM-YYYY}}

\item{end}{end date \code{MM-YYYY}}

\item{gr}{\code{monthly, weekly} or \code{daily}}

\item{md}{Defaults to False, set to True if you want main domain only}
}
\value{
The \code{curl} object with \code{status_code, headers, content} etc.
}
\description{
Get domain visits
}
\examples{
\dontrun{
get_visits(api_key = "my_api_key",
           domain = "cnn.com",
           start = "01-2016",
           end = "03-2016",
           gr = "monthly")
}
}
